refactor name of elasticsearch
